Update "Getting Started" and "Power" pages
This change adds new content and links to the "Getting Started" page and
fixes a broken link. Minor changes have been made to some of
the newly-linked pages.

I also updated the "Power" page based on my personal notes on the
Screeps source code. I ended up rewriting a significant portion of the
article while trying to work in the new information.